Models, Models, and more Models!
Updated with a GitHub Plugin for Unity (might not be worth it, but could be helpful too.)  Also added the building model with the start of colliders on the walls and floors, a temp_player object with broken script attached.  It's too late and I'm too brain dead to fix it right now, but figured I'd push what I have so far to my branch.